+++ CHANGES.md
@@ -8,6 +8,9 @@
 
 -   Always upload external triggers to the 'External' group.
 
+-   Add rudimentary burst event workflow to orchestrator: it just generates
+    VOEvents and circulars.
+
 ## 0.0.14 (2017-06-28)
 
 -   Fix typo that was causing a task to fail.